fix: withConn should not use a select to detect context cancellation
The function passed to withConn is already context aware. By using a separate select we only add another level of indirection which makes it hard to analyze the program and provides less-than-useful error messages in case the context is cancelled. Additionally, we now have a race condition between the two select cases when the context is cancelled leading to possibly inconsistent error messages. In reality the case for the local context seems to win out almost all the time which only exacerbates the problem of meaningless error messages.

Signed-off-by: Knut Zuidema <me@knutzuidema.de>